feat: add recursive copying

+++ 04-copy-directory/index.js
@@ -3,14 +3,21 @@ const fs = require('fs')
 const dirSorce = `${__dirname}/files`
 const dir = `${__dirname}/files-copy`
 
-async function copyFiles(){
+async function copyFiles(dirSorce, dir){
   await fs.promises.mkdir(dir, { recursive: true })
   const files = await fs.promises.readdir(dirSorce,{withFileTypes: true});
   for (const file of files){
-    await fs.promises.copyFile(`${dirSorce}/${file.name}`,`${dir}/${file.name}`)
-    console.log(`${dirSorce}/${file.name}`)
+    if(file.isFile()){
+      await fs.promises.copyFile(`${dirSorce}/${file.name}`,`${dir}/${file.name}`)
+      console.log(`${dirSorce}/${file.name}`)
+    }else{
+      await copyFiles(`${dirSorce}/${file.name}`,`${dir}/${file.name}`)
+    }
+    
   }
 }
+
+
 async function deleteFolder(){
   try{
     await fs.promises.access(dir)
@@ -22,7 +29,7 @@ async function deleteFolder(){
 
 async function copyDirectory(){
   await deleteFolder()
-  await copyFiles()
+  await copyFiles(dirSorce, dir)
 }
 
 copyDirectory()
